integrated account selection from wallet panel
