+++ BraveWallet/WalletHostingViewController.swift
@@ -27,6 +27,8 @@ public enum PresentingContext {
   case requestEthererumPermissions(_ request: WebpagePermissionRequest)
   /// Shows the user only the unlock/setup screen then dismisses to view an unlocked panel
   case panelUnlockOrSetup
+  /// Shows the user available wallet accounts to use
+  case accountSelection
 }
 
 /// The initial wallet controller to present when the user wants to view their wallet
